# api/openapi.yaml
openapi: 3.0.3
info:
  title: Our Sell API
  version: 1.0.0
  description: |
    Secure, versioned authentication and session-management API. Refresh
    tokens are opaque, rotated, and stored only as SHA-256 hashes on the server.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/auth/facebook/connect:
    post:
      operationId: startFacebookConnect
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Facebook authorization URL. The URL is safe to redirect the current browser to.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/FacebookConnectResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /api/v1/auth/facebook/callback:
    get:
      operationId: facebookCallback
      security: []
      description: Meta redirects here. The endpoint consumes the one-time state and redirects back to the fixed admin frontend URL.
      parameters:
        - { name: code, in: query, schema: { type: string } }
        - { name: state, in: query, schema: { type: string } }
        - { name: error, in: query, schema: { type: string } }
        - { name: error_reason, in: query, schema: { type: string } }
      responses:
        '302': { description: Redirects to the admin Facebook accounts view. }
  /api/v1/facebook/connections:
    get:
      operationId: listFacebookConnections
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Safe metadata for active Facebook connections owned by the current user.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/FacebookConnection' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/facebook/pages:
    get:
      operationId: listFacebookPages
      security: [{ bearerAuth: [] }]
      parameters:
        - name: connection_id
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Pages returned by Meta for the selected owned connection.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/FacebookPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/facebook/page-inventory:
    get:
      operationId: listFacebookPageInventory
      security: [{ bearerAuth: [] }]
      parameters:
        - name: connection_id
          in: query
          schema: { type: string, format: uuid }
        - name: search
          in: query
          schema: { type: string, maxLength: 100 }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
      responses:
        '200':
          description: Stored Facebook Page inventory belonging to the current user. Requires administrator access.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/FacebookPageInventory' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/facebook/pages/{pageID}/listing:
    put:
      operationId: saveFacebookPageListing
      security: [{ bearerAuth: [] }]
      parameters:
        - name: pageID
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FacebookListingInput' }
      responses:
        '200':
          description: Saved marketplace listing.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/FacebookPageListing' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/facebook/connections/{connectionID}/sync:
    post:
      operationId: syncFacebookConnection
      security: [{ bearerAuth: [] }]
      parameters:
        - name: connectionID
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Facebook Pages synchronized into the inventory.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [page_count]
                    properties:
                      page_count: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/marketplace/pages:
    get:
      operationId: listPublishedFacebookPageListings
      security: []
      parameters:
        - name: search
          in: query
          schema: { type: string, maxLength: 100 }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
      responses:
        '200':
          description: Published Facebook Page marketplace listings.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/FacebookPageInventory' }
  /api/v1/marketplace/pages/{listingID}:
    get:
      operationId: getPublishedFacebookPageListing
      security: []
      parameters:
        - name: listingID
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Published Facebook Page marketplace listing.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/FacebookPublicListing' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/marketplace/pages/{listingID}/orders:
    post:
      operationId: createMarketplaceOrder
      security: [{ bearerAuth: [] }]
      parameters:
        - name: listingID
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MarketplaceOrderInput' }
      responses:
        '201':
          description: Listing reservation and order created. Wallet orders are credited from an available balance hold atomically.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/MarketplaceOrder' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409':
          description: The listing is already reserved, sold, or unavailable.
          content:
              application/json:
                schema: { $ref: '#/components/schemas/Error' }
...
  /api/v1/facebook/connections/{connectionID}:
    delete:
      operationId: disconnectFacebook
      security: [{ bearerAuth: [] }]
      parameters:
        - name: connectionID
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { $ref: '#/components/responses/Message' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '400': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Message:
      description: Operation result.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Message' }
    ValidationError:
      description: Invalid request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: VALIDATION_ERROR
              message: Invalid request
              fields: { email: Invalid email address }
...
    Unauthorized:
      description: Missing, invalid, expired, or revoked credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but not authorized.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Resource already exists.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    DependencyUnavailable:
      description: Required dependency is unavailable.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    FacebookConnectResponse:
      type: object
      required: [authorization_url]
      properties:
        authorization_url: { type: string, format: uri }
    FacebookConnection:
      type: object
      required: [id, facebook_account_id, name, connected_at]
      properties:
        id: { type: string, format: uuid }
        facebook_account_id: { type: string }
        name: { type: string }
        connected_at: { type: string, format: date-time }
        token_expires_at: { type: string, format: date-time, nullable: true }
    FacebookPage:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
    FacebookListingInput:
      type: object
      required: [title, subtitle, description, price_usd, currency, delivery_window, status, featured, sort_order]
      additionalProperties: false
      properties:
        title: { type: string, minLength: 1, maxLength: 255 }
        subtitle: { type: string, minLength: 1, maxLength: 255 }
        description: { type: string, maxLength: 5000 }
        price_usd: { type: number, format: double, minimum: 0, example: 6.00 }
        currency: { type: string, enum: [USD], example: USD }
        delivery_window: { type: string, minLength: 1, maxLength: 120 }
        status: { type: string, enum: [draft, pending_review, published, unlisted] }
        featured: { type: boolean }
        sort_order: { type: integer, minimum: 0 }
    FacebookPageListing:
      allOf:
        - $ref: '#/components/schemas/FacebookListingInput'
        - type: object
          required: [id, page_id, created_at, updated_at]
          properties:
            id: { type: string, format: uuid }
            page_id: { type: string, format: uuid }
            published_at: { type: string, format: date-time, nullable: true }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    FacebookPageInventoryItem:
      type: object
      required: [id, facebook_page_id, name, connection_id, connection_name]
      properties:
        id: { type: string, format: uuid }
        facebook_page_id: { type: string }
        name: { type: string }
        connection_id: { type: string, format: uuid }
        connection_name: { type: string }
        last_synced_at: { type: string, format: date-time, nullable: true }
        listing: { $ref: '#/components/schemas/FacebookPageListing' }
    FacebookPageInventory:
      type: object
      required: [pages, total, page, page_size, total_pages]
      properties:
        pages:
          type: array
          items: { $ref: '#/components/schemas/FacebookPageInventoryItem' }
        total: { type: integer, format: int64 }
        page: { type: integer }
        page_size: { type: integer }
        total_pages: { type: integer }
    FacebookPublicListing:
      allOf:
        - $ref: '#/components/schemas/FacebookPageListing'
        - type: object
          required: [facebook_page_id, page_name]
          properties:
            facebook_page_id: { type: string }
            page_name: { type: string }
    MarketplaceOrderInput:
      type: object
      required: [buyer_note, payment_method, terms_accepted]
      additionalProperties: false
      properties:
        buyer_note: { type: string, maxLength: 2000 }
        payment_method: { type: string, enum: [wallet] }
        terms_accepted: { type: boolean, enum: [true] }
    MarketplaceOrder:
      type: object
      required: [id, reference, listing_id, buyer_user_id, amount_usd, currency, status, reservation_expires_at, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        reference: { type: string }
        listing_id: { type: string, format: uuid }
        buyer_user_id: { type: string, format: uuid }
        buyer_note: { type: string }
        amount_usd: { type: number, format: double, minimum: 0, example: 6.00 }
        currency: { type: string, enum: [USD] }
        payment_method: { type: string, enum: [wallet, bakong, card, usdt] }
        status: { type: string, enum: [pending_payment, payment_failed, paid, transfer_pending, completed, cancelled, expired] }
        reservation_expires_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        paid_at: { type: string, format: date-time, nullable: true }
        completed_at: { type: string, format: date-time, nullable: true }
...
    Message:
      type: object
      required: [message]
      properties:
        message: { type: string }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            fields: { type: object, additionalProperties: { type: string } }